+++ web/jobs/assets/BrowserView-BgsnOSEH.js
@@ -1,4 +1,4 @@
-import{r,j as e}from"./react-vendor-ol50NMz1.js";import{i as X,r as Z,C as ee,E as se}from"./application-flow-B3rFfdfn.js";import{c as j,z as ie,E as ne,e as w,t as ae,T as le,G as re,C as $,A as x,H as ce,D as g}from"./index-DEl2N-v9.js";import{P as B}from"./play-COSlRmth.js";import{M as D}from"./monitor-up-CcJtVx_Z.js";/**
+import{r,j as e}from"./react-vendor-ol50NMz1.js";import{i as X,r as Z,C as ee,E as se}from"./application-flow-D9ukx2CN.js";import{c as j,z as ie,E as ne,e as w,t as ae,T as le,G as re,C as $,A as x,H as ce,D as g}from"./index-D59JGUzx.js";import{P as B}from"./play-wg5rgG1v.js";import{M as D}from"./monitor-up-CvBcOWIY.js";/**
  * @license lucide-react v0.500.0 - ISC
  *
  * This source code is licensed under the ISC license.